+++ core/src/main/kotlin/net/corda/core/flows/ReceiveTransactionFlow.kt
@@ -23,9 +23,13 @@ import java.security.SignatureException
  * @property checkSufficientSignatures if true checks all required signatures are present. See [SignedTransaction.verify].
  * @property statesToRecord which transaction states should be recorded in the vault, if any.
  */
-class ReceiveTransactionFlow @JvmOverloads constructor(private val otherSideSession: FlowSession,
-                                                       private val checkSufficientSignatures: Boolean = true,
-                                                       private val statesToRecord: StatesToRecord = StatesToRecord.NONE) : FlowLogic<SignedTransaction>() {
+open class ReceiveTransactionFlow @JvmOverloads constructor(private val otherSideSession: FlowSession,
+                                                            private val checkSufficientSignatures: Boolean = true,
+                                                            private val statesToRecord: StatesToRecord = StatesToRecord.NONE) : FlowLogic<SignedTransaction>() {
+    // TODO Should this be public API?
+    // Capture the received tx in case we need it even if we deem it invalid
+    private var _receivedTransaction: SignedTransaction? = null
+
     @Suppress("KDocMissingDocumentation")
     @Suspendable
     @Throws(SignatureException::class,
@@ -39,8 +43,9 @@ class ReceiveTransactionFlow @JvmOverloads constructor(private val otherSideSess
             logger.trace { "Receiving a transaction (but without checking the signatures) from ${otherSideSession.counterparty}" }
         }
         val stx = otherSideSession.receive<SignedTransaction>().unwrap {
+            _receivedTransaction = it
             it.pushToLoggingContext()
-            logger.info("Received transaction acknowledgement request from party ${otherSideSession.counterparty.name}.")
+            logger.info("Received transaction acknowledgement request from party ${otherSideSession.counterparty}.")
             subFlow(ResolveTransactionsFlow(it, otherSideSession))
             logger.info("Transaction dependencies resolution completed.")
             try {
@@ -54,12 +59,21 @@ class ReceiveTransactionFlow @JvmOverloads constructor(private val otherSideSess
         if (checkSufficientSignatures) {
             // We should only send a transaction to the vault for processing if we did in fact fully verify it, and
             // there are no missing signatures. We don't want partly signed stuff in the vault.
+            checkBeforeRecording(stx)
             logger.info("Successfully received fully signed tx. Sending it to the vault for processing.")
             serviceHub.recordTransactions(statesToRecord, setOf(stx))
             logger.info("Successfully recorded received transaction locally.")
         }
         return stx
     }
+
+    /**
+     * Hook to perform extra checks on the received transaction just before it's recorded. The transaction has already
+     * been resolved and verified at this point.
+     */
+    @Suspendable
+    @Throws(FlowException::class)
+    protected open fun checkBeforeRecording(stx: SignedTransaction) = Unit
 }
 
 /**