+++ docs/source/example-code/src/integration-test/java/net/corda/docs/JavaIntegrationTestingTutorial.java
@@ -18,7 +18,9 @@ import net.corda.testing.node.User;
 import org.junit.Test;
 import rx.Observable;
 
-import java.util.*;
+import java.util.Currency;
+import java.util.HashSet;
+import java.util.List;
 
 import static java.util.Arrays.asList;
 import static java.util.Collections.singletonList;
@@ -38,7 +40,7 @@ public class JavaIntegrationTestingTutorial {
         // START 1
         driver(new DriverParameters()
                 .withStartNodesInProcess(true)
-                .withExtraCordappPackagesToScan(Arrays.asList("net.corda.finance.contracts.asset", "net.corda.finance.schemas")), dsl -> {
+                .withExtraCordappPackagesToScan(singletonList("net.corda.finance")), dsl -> {
 
             User aliceUser = new User("aliceUser", "testPassword1", new HashSet<>(asList(
                     startFlow(CashIssueAndPaymentFlow.class),