+++ docs/source/changelog.rst
@@ -7,9 +7,17 @@ release, see :doc:`upgrade-notes`.
 Unreleased
 ----------
 
-* Introduced new optional network bootstrapper command line option (--minimum-platform-version) to set as a network parameter
+* CorDapps now have the ability to specify a minimum platform version in their MANIFEST.MF to prevent old nodes from loading them.
 
-* Introduce minimum and target platform version for CorDapps.
+* CorDapps have the ability to specify a target platform version in their MANIFEST.MF as a means of indicating to the node
+  the app was designed and tested on that version.
+
+* ``FinalityFlow`` is now an inlined flow and no longer requires a handler flow in the counterparty. This is to fix the
+  security problem with the handler flow as it accepts any transaction it receives without any checks. Existing CorDapp
+  binaries relying on this old behaviour will continue to function as previously. However, it is strongly recommended that
+  CorDapps switch to this new API. See :doc:`upgrade-notes` for further details.
+
+* Introduced new optional network bootstrapper command line option (--minimum-platform-version) to set as a network parameter
 
 * BFT-Smart and Raft notary implementations have been extracted out of node into ``experimental`` CorDapps to emphasise
   their experimental nature. Moreover, the BFT-Smart notary will only work in dev mode due to its use of Java serialization.